Add verification if user is the user who wants to make action

+++ authentication/views.py
@@ -46,8 +46,16 @@ class UserViewset(ModelViewSet):
         serializer = self.get_serializer(product)
         return Response(serializer.data)
 
+    def check_permission(self, product, user):
+        if product != user:
+            return Response({'detail': 'Vous ne pouvez modifier que votre profil'}, status=status.HTTP_401_UNAUTHORIZED)
+        return None
+
     def update(self, request, pk=None):
         product = self.get_object()
+        permission_error = self.check_permission(product, request.user)
+        if permission_error:
+            return permission_error
         serializer = self.get_serializer(product, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -57,6 +65,9 @@ class UserViewset(ModelViewSet):
 
     def partial_update(self, request, pk=None):
         product = self.get_object()
+        permission_error = self.check_permission(product, request.user)
+        if permission_error:
+            return permission_error
         serializer = self.get_serializer(product, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -66,5 +77,8 @@ class UserViewset(ModelViewSet):
 
     def destroy(self, request, pk=None):
         product = self.get_object()
+        permission_error = self.check_permission(product, request.user)
+        if permission_error:
+            return permission_error
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
